updating user info on generate meal plan

+++ src/main/java/com/example/prepitbackend/domain/User.java
@@ -42,8 +42,8 @@ public class User implements UserDetails{
     private String email;
     private String firstName;
     private String lastName;
-    private int weight;
-    private int height;
+    private Double weight;
+    private Double height;
     private int age;
     private String activityType;
     private String gender;
